Don't crash on non-ASCII characters in OCR output
When saving to PDF, non-ASCII characters in the OCR output could crash the
worker thread due to an uncaught exception thrown by Encode::decode.  This
happened because we were trying to decode a string that was already decoded.

The test case that seemed to indicate that a decode() was missing, was in fact
missing a 'use utf8' itself.

While we're at it, also remove another couple of unneeded utf8-related things.